+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Add `String#allIndexesOf(needle)` to get an array of indexes
 * Speed up `String#tokenizeHTML(source)` when supplying custom blocks
+* Add `Function#enforceProperty(setter)`
 
 ## 0.6.5 (2019-02-18)